Repo Gardening: consolidate existing triage tasks into one
Instead of having 2 tasks (`notifyKitKat` and `triageNewIssues`) that both look at issues and apply labels / send Slack notifications based on priority and other factors, let's have only one task.

This has the added benefit of ensuring that triaging tasks get triggered every time, when an issue is opened but also when labels are added to an existing issue.

Since we're removing 2 existing tasks, this will be a major bump for this GitHub action. However, we are not losing any functionality, as long as you replace active tasks.
