+++ .github/workflows/gardening.yml
@@ -78,4 +78,3 @@ jobs:
           slack_editorial_channel: ${{ secrets.SLACK_EDITORIAL_CHANNEL }}
           slack_he_triage_channel: ${{ secrets.SLACK_HE_TRIAGE_CHANNEL }}
           slack_quality_channel: ${{ secrets.SLACK_QUALITY_CHANNEL }}
-          slack_kitkat_channel: ${{ secrets.SLACK_KITKAT_CHANNEL }}
