+++ projects/github-actions/repo-gardening/action.yml
@@ -36,10 +36,6 @@ inputs:
     description: "Slack channel ID where issues needing extra triage / escalation will be sent"
     required: false
     default: ""
-  slack_kitkat_channel:
-    description: "Slack channel ID where issues needing the attention of the WordPress.com Quality team will be sent"
-    required: false
-    default: ""
   tasks:
     description: "Comma-separated selection of task names (function name, camelCase) this action should run. e.g. addLabels,cleanLabels"
     required: false
