+++ projects/github-actions/repo-gardening/changelog/update-repo-gardening-triage-tasks
@@ -1,0 +1,4 @@
+Significance: major
+Type: removed
+
+The notifyKitKat and triageNewIssues tasks have been removed. The new triageIssues task now handles both of those tasks.
