+++ projects/github-actions/repo-gardening/changelog/update-repo-gardening-triage-tasks#2
@@ -1,0 +1,4 @@
+Significance: minor
+Type: added
+
+Create a new triageIssues task, to handle automated label triage and Slack notifications for important issues.
