+++ projects/github-actions/repo-gardening/src/index.js
@@ -9,9 +9,8 @@ const flagOss = require( './tasks/flag-oss' );
 const gatherSupportReferences = require( './tasks/gather-support-references' );
 const notifyDesign = require( './tasks/notify-design' );
 const notifyEditorial = require( './tasks/notify-editorial' );
-const notifyKitKat = require( './tasks/notify-kitkat' );
 const replyToCustomersReminder = require( './tasks/reply-to-customers-reminder' );
-const triageNewIssues = require( './tasks/triage-new-issues' );
+const triageIssues = require( './tasks/triage-issues' );
 const wpcomCommitReminder = require( './tasks/wpcom-commit-reminder' );
 const debug = require( './utils/debug' );
 const ifNotClosed = require( './utils/if-not-closed' );
@@ -54,11 +53,6 @@ const automations = [
 		task: ifNotClosed( notifyEditorial ),
 	},
 	{
-		event: 'issues',
-		action: [ 'labeled' ],
-		task: notifyKitKat,
-	},
-	{
 		event: 'push',
 		task: wpcomCommitReminder,
 	},
@@ -69,8 +63,8 @@ const automations = [
 	},
 	{
 		event: 'issues',
-		action: [ 'opened', 'reopened' ],
-		task: triageNewIssues,
+		action: [ 'opened', 'reopened', 'labeled' ],
+		task: triageIssues,
 	},
 	{
 		event: 'issues',
