--- projects/github-actions/repo-gardening/src/tasks/notify-kitkat/readme.md
@@ -1,16 +1,0 @@
-# Notify KitKat
-
-**Kitkat is the WordPress.com Quality team. If your quality team is named differently, you can still use this task! :)**
-
-Send a Slack notification to KitKat when there is a high or blocker priority bug, or a bug with no priority label.
-
-This task responds to the following labels:
-
-- "[Pri] BLOCKER"
-- "[Pri] High"
-
-When one of the labels above is added to an issue, a Slack message will be sent to channel of your choice.
-
-## Rationale
-
-This alert will help improve triage times which should then improve MTTR.
