--- projects/github-actions/repo-gardening/src/tasks/triage-new-issues/readme.md
@@ -1,13 +1,0 @@
-# Triage New Issues
-
-Add labels to newly opened issues to help triage the issue:
-
-1. Add a `[Plugin]` label if a specific label is impacted by the issue.
-2. Add a `[Pri]` label if we can determine the severity of the issue from the issue contents.
-3. Add a `[Platform]` label if the issue is specific to a platform.
-
-This is best used in combination with GitHub [issue forms](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository#creating-issue-forms).
-
-## Rationale
-
-Adding a label to new issues allows each team to get quick notices of issues that need their attention, without having to wait for manual triage. Priority also allows them to sort through the issues in the right order.
